requirements.txt
asgiref==3.3.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
dj-database-url==0.5.0
Django==3.1.7
django-cors-headers==3.7.0
django-environ==0.4.5
django-extensions==3.1.1
django-heroku==0.3.1
djangorestframework==3.12.2
drf-jwt==1.19.0
environ==1.0
flake8==3.9.0
gunicorn==20.1.0
idna==2.10
importlib-metadata==3.7.3
mccabe==0.6.1
psycopg2==2.9.3
pycodestyle==2.7.0
pycparser==2.20
pyflakes==2.3.0
PyJWT==1.7.1
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
typing-extensions==3.7.4.3
urllib3==1.26.4
whitenoise==5.2.0
zipp==3.4.1
